+++ src/main/java/gr/hua/distributed/it21774/service/AppUserDetailsImpl.java
@@ -30,9 +30,9 @@ public class AppUserDetailsImpl implements UserDetails {
 
     private String lastName;
 
-    private int afm;
+    private Long afm;
 
-    private int amka;
+    private Long amka;
 
     // Constructor
 
@@ -43,8 +43,8 @@ public class AppUserDetailsImpl implements UserDetails {
                               String email,
                               String firstName,
                               String lastName,
-                              int afm,
-                              int amka) {
+                              Long afm,
+                              Long amka) {
         this.id = id;
         this.username = username;
         this.password = password;
@@ -106,11 +106,11 @@ public class AppUserDetailsImpl implements UserDetails {
         return lastName;
     }
 
-    public int getAfm() {
+    public Long getAfm() {
         return afm;
     }
 
-    public int getAmka() {
+    public Long getAmka() {
         return amka;
     }
